Cut and paste StatefulEffect to new files
