+++ src/effects/AutoDuck.h
@@ -11,7 +11,7 @@
 #ifndef __AUDACITY_EFFECT_AUTODUCK__
 #define __AUDACITY_EFFECT_AUTODUCK__
 
-#include "Effect.h"
+#include "StatefulEffect.h"
 #include "../ShuttleAutomation.h"
 #include <float.h> // for DBL_MAX
 #include "../widgets/wxPanelWrapper.h"
