+++ src/effects/Effect.cpp
@@ -43,23 +43,6 @@ static_assert(kPlayID == EffectEditor::kPlayID);
 
 using t2bHash = std::unordered_map< void*, bool >;
 
-bool StatefulEffect::Instance::Process(EffectSettings &settings)
-{
-   return GetEffect().Process(*this, settings);
-}
-
-auto StatefulEffect::Instance::GetLatency(const EffectSettings &, double) const
-   -> SampleCount
-{
-   return GetEffect().GetLatency().as_long_long();
-}
-
-size_t StatefulEffect::Instance::ProcessBlock(EffectSettings &,
-   const float *const *, float *const *, size_t)
-{
-   return 0;
-}
-
 Effect::Effect()
 {
 }
@@ -127,49 +110,6 @@ auto Effect::RealtimeSupport() const -> RealtimeSince
 bool Effect::SupportsAutomation() const
 {
    return true;
-}
-
-StatefulEffect::~StatefulEffect() = default;
-
-std::shared_ptr<EffectInstance> StatefulEffect::MakeInstance() const
-{
-   // Cheat with const-cast to return an object that calls through to
-   // non-const methods of a stateful effect.
-   // Stateless effects should override this function and be really const
-   // correct.
-   return std::make_shared<Instance>(const_cast<StatefulEffect&>(*this));
-}
-
-std::unique_ptr<EffectEditor> StatefulEffect::PopulateUI(
-   const EffectPlugin &, ShuttleGui &S,
-   EffectInstance &instance, EffectSettingsAccess &access,
-   const EffectOutputs *pOutputs) const
-{
-   auto parent = S.GetParent();
-
-   // As in MakeInstance, we still cheat const for stateful effects!
-   auto pThis = const_cast<StatefulEffect*>(this);
-
-   // Let the effect subclass provide its own editor if it wants
-   auto result = pThis->PopulateOrExchange(S, instance, access, pOutputs);
-
-   parent->SetMinSize(parent->GetSizer()->GetMinSize());
-
-   if (!result) {
-      // No custom editor object?  Then use the default
-      result = std::make_unique<DefaultEffectEditor>(*pThis,
-         *pThis, access, S.GetParent());
-      parent->PushEventHandler(pThis);
-   }
-   return result;
-}
-
-std::unique_ptr<EffectEditor> StatefulEffect::MakeEditor(
-   ShuttleGui &, EffectInstance &, EffectSettingsAccess &,
-   const EffectOutputs *) const
-{
-   assert(false);
-   return nullptr;
 }
 
 const EffectParameterMethods &Effect::Parameters() const
