+++ src/effects/Effect.h
@@ -14,8 +14,6 @@
 
 #include "EffectBase.h"
 
-#include "StatefulEffectBase.h"
-
 #define BUILTIN_EFFECT_PREFIX wxT("Built-in Effect: ")
 
 class EffectParameterMethods;
@@ -237,41 +235,6 @@ public:
    FetchParameters(Base &, EffectSettings &s) {
       return &GetSettings(s);
    }
-};
-
-//! Subclass of Effect, to be eliminated after all of its subclasses
-//! are rewritten to be stateless
-class StatefulEffect
-   : public StatefulEffectBase
-   , public Effect
-{
-public:
-   class AUDACITY_DLL_API Instance : public StatefulEffectBase::Instance {
-   public:
-      using StatefulEffectBase::Instance::Instance;
-      bool Process(EffectSettings &settings) override;
-      SampleCount GetLatency(
-         const EffectSettings &settings, double sampleRate) const override;
-      //! Default implementation fails (returns 0 always)
-      size_t ProcessBlock(EffectSettings &settings,
-         const float *const *inBlock, float *const *outBlock, size_t blockLen)
-      override;
-   };
-
-   ~StatefulEffect() override;
-
-   std::shared_ptr<EffectInstance> MakeInstance() const override;
-
-   //! Allows PopulateOrExchange to return null
-   std::unique_ptr<EffectEditor> PopulateUI(const EffectPlugin &plugin,
-      ShuttleGui &S, EffectInstance &instance, EffectSettingsAccess &access,
-      const EffectOutputs *pOutputs) const override;
-
-private:
-   //! Needed to make subclasses concrete, but should never be called
-   virtual std::unique_ptr<EffectEditor> MakeEditor(
-      ShuttleGui & S, EffectInstance &instance, EffectSettingsAccess &access,
-      const EffectOutputs *pOutputs) const final;
 };
 
 // FIXME:
