+++ src/effects/NoiseRemoval.h
@@ -16,7 +16,7 @@
 
 #if !defined(EXPERIMENTAL_NOISE_REDUCTION)
 
-#include "Effect.h"
+#include "StatefulEffect.h"
 #include "EffectUI.h"
 
 class wxButton;
