+++ src/effects/SoundTouchEffect.h
@@ -19,7 +19,7 @@
 #ifndef __AUDACITY_EFFECT_SOUNDTOUCH__
 #define __AUDACITY_EFFECT_SOUNDTOUCH__
 
-#include "Effect.h"
+#include "StatefulEffect.h"
 
 // forward declaration of a class defined in SoundTouch.h
 // which is not included here
